Move more config/setup into renin proper
